Fix cast to int

+++ src/main/java/com/sweetrpg/crafttracker/common/util/RecipeUtil.java
@@ -193,7 +193,7 @@ public class RecipeUtil {
                 }
 
                 if(cost > highestCost) {
-                    highestCost = cost;
+                    highestCost = (int) cost;
                 }
             }
         }
